# crl-release-20.2: db: mergingIter.SeekPrefixGE stops early if prefix cannot match