This is a 20.2 backport of #1085. While investigating cockroachlabs/support#898,
one node hit #1070:

<img width="1766" alt="Screen Shot 2021-03-25 at 5 33 13 PM" src="https://user-images.githubusercontent.com/867352/112658351-b49e2b00-8e29-11eb-8c83-4cf2d1f17263.png">

The cherrypick didn't apply cleanly because of other recent iterator optimizations,
so I think we should probably wait for @sumeerbhola to give this a review before
merging.

```
name                                                            old time/op  new time/op  delta
IteratorSeqSeekPrefixGENotFound/skip=1/with-tombstone=false-16  1.42µs ± 4%  1.36µs ± 6%     ~     (p=0.095 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=1/with-tombstone=true-16   11.0ms ± 5%   0.0ms ± 6%  -99.99%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/with-tombstone=false-16  1.39µs ± 4%  1.36µs ±11%     ~     (p=0.310 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=2/with-tombstone=true-16   11.2ms ± 5%   0.0ms ± 4%  -99.99%  (p=0.008 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/with-tombstone=false-16  1.38µs ± 4%  1.38µs ± 7%     ~     (p=1.000 n=5+5)
IteratorSeqSeekPrefixGENotFound/skip=4/with-tombstone=true-16   11.1ms ± 3%   0.0ms ± 9%  -99.99%  (p=0.008 n=5+5)
```